+++ xarray/core/indexing.py
@@ -2,7 +2,7 @@ import enum
 import functools
 import operator
 from collections import defaultdict
-from typing import Any, Callable, Iterable, List, Sequence, Tuple, Union
+from typing import Any, Callable, Iterable, List, Tuple, Union
 
 import numpy as np
 import pandas as pd
@@ -15,7 +15,7 @@ from .pycompat import (
     is_duck_dask_array,
     sparse_array_type,
 )
-from .utils import is_dict_like, maybe_cast_to_coords_dtype
+from .utils import maybe_cast_to_coords_dtype
 
 
 def expanded_indexer(key, ndim):
@@ -52,196 +52,44 @@ def _expand_slice(slice_, size):
     return np.arange(*slice_.indices(size))
 
 
-def _sanitize_slice_element(x):
-    from .dataarray import DataArray
-    from .variable import Variable
+def group_indexers_by_index(data_obj, indexers, method=None, tolerance=None):
+    # TODO: benbovy - flexible indexes: indexers are still grouped by dimension
+    # - Make xarray.Index hashable so that it can be used as key in a mapping?
+    indexes = {}
+    grouped_indexers = defaultdict(dict)
 
-    if isinstance(x, (Variable, DataArray)):
-        x = x.values
+    # TODO: data_obj.xindexes should eventually return the PandasIndex instance
+    # for each multi-index levels
+    xindexes = dict(data_obj.xindexes)
+    for level, dim in data_obj._level_coords.items():
+        xindexes[level] = xindexes[dim]
 
-    if isinstance(x, np.ndarray):
-        if x.ndim != 0:
-            raise ValueError(
-                f"cannot use non-scalar arrays in a slice for xarray indexing: {x}"
-            )
-        x = x[()]
-
-    return x
-
-
-def _asarray_tuplesafe(values):
-    """
-    Convert values into a numpy array of at most 1-dimension, while preserving
-    tuples.
-
-    Adapted from pandas.core.common._asarray_tuplesafe
-    """
-    if isinstance(values, tuple):
-        result = utils.to_0d_object_array(values)
-    else:
-        result = np.asarray(values)
-        if result.ndim == 2:
-            result = np.empty(len(values), dtype=object)
-            result[:] = values
-
-    return result
-
-
-def _is_nested_tuple(possible_tuple):
-    return isinstance(possible_tuple, tuple) and any(
-        isinstance(value, (tuple, list, slice)) for value in possible_tuple
-    )
-
-
-def get_indexer_nd(index, labels, method=None, tolerance=None):
-    """Wrapper around :meth:`pandas.Index.get_indexer` supporting n-dimensional
-    labels
-    """
-    flat_labels = np.ravel(labels)
-    flat_indexer = index.get_indexer(flat_labels, method=method, tolerance=tolerance)
-    indexer = flat_indexer.reshape(labels.shape)
-    return indexer
-
-
-def convert_label_indexer(index, label, index_name="", method=None, tolerance=None):
-    """Given a pandas.Index and labels (e.g., from __getitem__) for one
-    dimension, return an indexer suitable for indexing an ndarray along that
-    dimension. If `index` is a pandas.MultiIndex and depending on `label`,
-    return a new pandas.Index or pandas.MultiIndex (otherwise return None).
-    """
-    from .indexes import PandasIndex
-
-    new_index = None
-
-    if isinstance(label, slice):
-        if method is not None or tolerance is not None:
-            raise NotImplementedError(
-                "cannot use ``method`` argument if any indexers are slice objects"
-            )
-        indexer = index.slice_indexer(
-            _sanitize_slice_element(label.start),
-            _sanitize_slice_element(label.stop),
-            _sanitize_slice_element(label.step),
-        )
-        if not isinstance(indexer, slice):
-            # unlike pandas, in xarray we never want to silently convert a
-            # slice indexer into an array indexer
-            raise KeyError(
-                "cannot represent labeled-based slice indexer for dimension "
-                f"{index_name!r} with a slice over integer positions; the index is "
-                "unsorted or non-unique"
-            )
-
-    elif is_dict_like(label):
-        is_nested_vals = _is_nested_tuple(tuple(label.values()))
-        if not isinstance(index, pd.MultiIndex):
-            raise ValueError(
-                "cannot use a dict-like object for selection on "
-                "a dimension that does not have a MultiIndex"
-            )
-        elif len(label) == index.nlevels and not is_nested_vals:
-            indexer = index.get_loc(tuple(label[k] for k in index.names))
-        else:
-            for k, v in label.items():
-                # index should be an item (i.e. Hashable) not an array-like
-                if isinstance(v, Sequence) and not isinstance(v, str):
-                    raise ValueError(
-                        "Vectorized selection is not "
-                        "available along level variable: " + k
-                    )
-            indexer, new_index = index.get_loc_level(
-                tuple(label.values()), level=tuple(label.keys())
-            )
-
-            # GH2619. Raise a KeyError if nothing is chosen
-            if indexer.dtype.kind == "b" and indexer.sum() == 0:
-                raise KeyError(f"{label} not found")
-
-    elif isinstance(label, tuple) and isinstance(index, pd.MultiIndex):
-        if _is_nested_tuple(label):
-            indexer = index.get_locs(label)
-        elif len(label) == index.nlevels:
-            indexer = index.get_loc(label)
-        else:
-            indexer, new_index = index.get_loc_level(
-                label, level=list(range(len(label)))
-            )
-    else:
-        label = (
-            label
-            if getattr(label, "ndim", 1) > 1  # vectorized-indexing
-            else _asarray_tuplesafe(label)
-        )
-        if label.ndim == 0:
-            # see https://github.com/pydata/xarray/pull/4292 for details
-            label_value = label[()] if label.dtype.kind in "mM" else label.item()
-            if isinstance(index, pd.MultiIndex):
-                indexer, new_index = index.get_loc_level(label_value, level=0)
-            elif isinstance(index, pd.CategoricalIndex):
-                if method is not None:
-                    raise ValueError(
-                        "'method' is not a valid kwarg when indexing using a CategoricalIndex."
-                    )
-                if tolerance is not None:
-                    raise ValueError(
-                        "'tolerance' is not a valid kwarg when indexing using a CategoricalIndex."
-                    )
-                indexer = index.get_loc(label_value)
-            else:
-                indexer = index.get_loc(label_value, method=method, tolerance=tolerance)
-        elif label.dtype.kind == "b":
-            indexer = label
-        else:
-            if isinstance(index, pd.MultiIndex) and label.ndim > 1:
-                raise ValueError(
-                    "Vectorized selection is not available along "
-                    "MultiIndex variable: " + index_name
-                )
-            indexer = get_indexer_nd(index, label, method, tolerance)
-            if np.any(indexer < 0):
-                raise KeyError(f"not all values found in index {index_name!r}")
-
-    if new_index is not None:
-        new_index = PandasIndex(new_index)
-
-    return indexer, new_index
-
-
-def get_dim_indexers(data_obj, indexers):
-    """Given a xarray data object and label based indexers, return a mapping
-    of label indexers with only dimension names as keys.
-
-    It groups multiple level indexers given on a multi-index dimension
-    into a single, dictionary indexer for that dimension (Raise a ValueError
-    if it is not possible).
-    """
-    invalid = [
-        k
-        for k in indexers
-        if k not in data_obj.dims and k not in data_obj._level_coords
-    ]
-    if invalid:
-        raise ValueError(f"dimensions or multi-index levels {invalid!r} do not exist")
-
-    level_indexers = defaultdict(dict)
-    dim_indexers = {}
     for key, label in indexers.items():
-        (dim,) = data_obj[key].dims
-        if key != dim:
-            # assume here multi-index level indexer
-            level_indexers[dim][key] = label
-        else:
-            dim_indexers[key] = label
+        try:
+            index = xindexes[key]
+            coord = data_obj.coords[key]
+            dim = coord.dims[0]
+            if dim not in indexes:
+                indexes[dim] = index
 
-    for dim, level_labels in level_indexers.items():
-        if dim_indexers.get(dim, False):
-            raise ValueError(
-                "cannot combine multi-index level indexers with an indexer for "
-                f"dimension {dim}"
-            )
-        dim_indexers[dim] = level_labels
+            label = maybe_cast_to_coords_dtype(label, coord.dtype)
+            grouped_indexers[dim][key] = label
 
-    return dim_indexers
+        except KeyError:
+            if key in data_obj.coords:
+                raise KeyError(f"no index found for coordinate {key}")
+            elif key not in data_obj.dims:
+                raise KeyError(f"{key} is not a valid dimension or coordinate")
+            # key is a dimension without coordinate: we'll reuse the provided labels
+            elif method is not None or tolerance is not None:
+                raise ValueError(
+                    "cannot supply ``method`` or ``tolerance`` "
+                    "when the indexed dimension does not have "
+                    "an associated coordinate."
+                )
+            grouped_indexers[None][key] = label
+
+    return indexes, grouped_indexers
 
 
 def remap_label_indexers(data_obj, indexers, method=None, tolerance=None):
@@ -255,26 +103,25 @@ def remap_label_indexers(data_obj, indexers, method=None, tolerance=None):
     pos_indexers = {}
     new_indexes = {}
 
-    dim_indexers = get_dim_indexers(data_obj, indexers)
-    for dim, label in dim_indexers.items():
-        try:
-            index = data_obj.xindexes[dim].to_pandas_index()
-        except KeyError:
-            # no index for this dimension: reuse the provided labels
-            if method is not None or tolerance is not None:
-                raise ValueError(
-                    "cannot supply ``method`` or ``tolerance`` "
-                    "when the indexed dimension does not have "
-                    "an associated coordinate."
-                )
+    indexes, grouped_indexers = group_indexers_by_index(
+        data_obj, indexers, method, tolerance
+    )
+
+    forward_pos_indexers = grouped_indexers.pop(None, None)
+    if forward_pos_indexers is not None:
+        for dim, label in forward_pos_indexers.items():
             pos_indexers[dim] = label
-        else:
-            coords_dtype = data_obj.coords[dim].dtype
-            label = maybe_cast_to_coords_dtype(label, coords_dtype)
-            idxr, new_idx = convert_label_indexer(index, label, dim, method, tolerance)
-            pos_indexers[dim] = idxr
-            if new_idx is not None:
-                new_indexes[dim] = new_idx
+
+    for dim, index in indexes.items():
+        labels = grouped_indexers[dim]
+        idxr, new_idx = index.query(labels, method=method, tolerance=tolerance)
+        pos_indexers[dim] = idxr
+        if new_idx is not None:
+            new_indexes[dim] = new_idx
+
+    # TODO: benbovy - flexible indexes: support the following cases:
+    # - an index query returns positional indexers over multiple dimensions
+    # - check/combine positional indexers returned by multiple indexes over the same dimension
 
     return pos_indexers, new_indexes